move shapes messages into api folder
