+++ nanome/api/message_serializers.py
@@ -1,13 +1,9 @@
 import logging
-import types
 
 from nanome._internal import serializer_fields as serializer_fields
-
 from nanome._internal.enums import IntegrationCommands
-from nanome._internal.enum_utils import IntEnum
-from nanome.api import integration, structure, macro, shapes, ui, volumetric
+from nanome.api import integration, structure, volumetric
 from nanome.api._hashes import Hashes
-from nanome.api.streams import Stream
 from nanome.api.user import PresenterInfo
 
 
@@ -187,251 +183,6 @@ class SetSkybox(serializer_fields.TypeSerializer):
     def deserialize(self, version, context):
         raise NotImplementedError
 
-
-class DeleteShape(serializer_fields.TypeSerializer):
-
-    def __init__(self):
-        pass
-
-    def version(self):
-        return 1
-
-    def name(self):
-        return "DeleteShape"
-
-    def serialize(self, version, value, context):
-        if version == 0:
-            if len(value) > 1:
-                msg = "SetShape: Using a list of shapes with an old version of Nanome"
-                logger.warning(msg)
-                raise TypeError(msg)
-            context.write_int(value[0])
-        elif version == 1:
-            context.write_int_array(value)
-
-    def deserialize(self, version, context):
-        if version == 0:
-            return [context.read_bool()]
-        elif version == 1:
-            return context.read_byte_array()
-
-
-class SetShape(serializer_fields.TypeSerializer):
-
-    def __init__(self):
-        self._position = serializer_fields.UnityPositionSerializer()
-        self._rotation = serializer_fields.UnityRotationSerializer()
-        self._color = serializer_fields.ColorSerializer()
-        self._sphere = shapes.serializers.SphereSerializer()
-        self._line = shapes.serializers.LineSerializer()
-        self._label = shapes.serializers.LabelSerializer()
-        self._mesh = shapes.serializers.MeshSerializer()
-        self._shape = shapes.serializers.ShapeSerializer()
-        self._shape_array = serializer_fields.ArraySerializer()
-        self._shape_array.set_type(self._shape)
-
-    def version(self):
-        return 2
-
-    def name(self):
-        return "SetShape"
-
-    def serialize(self, version, value, context):
-        from nanome.util import Quaternion
-        from nanome.util.enums import ShapeType
-        if version == 0:
-            if len(value) > 1:
-                msg = "SetShape: Using a list of shapes with an old version of Nanome"
-                logger.warning(msg)
-                raise TypeError(msg)
-            first_elem = value[0]
-            context.write_byte(int(first_elem.shape_type))
-            if first_elem.shape_type == ShapeType.Sphere:
-                context.write_using_serializer(self._sphere, first_elem)
-            if first_elem.shape_type == ShapeType.Line:
-                context.write_using_serializer(self._line, first_elem)
-            if first_elem.shape_type == ShapeType.Label:
-                context.write_using_serializer(self._label, first_elem)
-            if first_elem.shape_type == ShapeType.Mesh:
-                context.write_using_serializer(self._mesh, first_elem)
-            context.write_int(first_elem.index)
-            context.write_long(first_elem.target)
-            context.write_byte(int(first_elem.anchor))
-            context.write_using_serializer(self._position, first_elem.position)
-            context.write_using_serializer(self._rotation, Quaternion())
-            context.write_using_serializer(self._color, first_elem.color)
-        elif version == 1:
-            if len(value) > 1:
-                msg = "SetShape: Using a list of shapes with an old version of Nanome"
-                logger.warning(msg)
-                raise TypeError(msg)
-            first_elem = value[0]
-            context.write_using_serializer(self._shape, first_elem)
-        elif version == 2:
-            context.write_using_serializer(self._shape_array, value)
-
-    def deserialize(self, version, context):
-        if version < 2:
-            return ([context.read_int()], [context.read_bool()])
-        else:
-            indices_arr = context.read_int_array()
-            success_arr = context.read_byte_array()
-            return (indices_arr, success_arr)
-
-
-class CreateStream(serializer_fields.TypeSerializer):
-
-    def __init__(self):
-        pass
-
-    def version(self):
-        return 2
-
-    def name(self):
-        return "StreamCreation"
-
-    def serialize(self, version, value, context):
-        from nanome.util.enums import StreamType as SType
-        stream_type = value[0]
-        if version > 0:
-            context.write_byte(stream_type)
-        if version >= 2:
-            context.write_byte(value[2])
-
-        if stream_type == SType.shape_position or stream_type == SType.shape_color or stream_type == SType.sphere_shape_radius:
-            context.write_int_array(value[1])
-        else:
-            context.write_long_array(value[1])
-
-    def deserialize(self, version, context):
-        raise NotImplementedError
-
-
-class CreateStreamResult(serializer_fields.TypeSerializer):
-
-    def __init__(self):
-        pass
-
-    def version(self):
-        return 2
-
-    def name(self):
-        return "StreamCreationResult"
-
-    def serialize(self, version, value, context):
-        raise NotImplementedError
-
-    def deserialize(self, version, context):
-        from nanome.util.enums import StreamDataType, StreamDirection
-        err = context.read_byte()
-        id = context.read_uint()
-        if version > 0:
-            data_type = StreamDataType(context.read_byte())
-        else:
-            data_type = StreamDataType.float
-        if version >= 2:
-            direction = StreamDirection(context.read_byte())
-        else:
-            direction = StreamDirection.writing
-        return (err, id, data_type, direction)
-
-
-class DestroyStream(serializer_fields.TypeSerializer):
-
-    def __init__(self):
-        pass
-
-    def version(self):
-        return 0
-
-    def name(self):
-        return "StreamDestruction"
-
-    def serialize(self, version, value, context):
-        context.write_uint(value)
-
-    def deserialize(self, version, context):
-        raise NotImplementedError
-
-
-class FeedStream(serializer_fields.TypeSerializer):
-
-    def __init__(self):
-        self.__array = serializer_fields.ArraySerializer()
-        self.__array.set_type(serializer_fields.StringSerializer())
-
-    def version(self):
-        return 2
-
-    def name(self):
-        return "StreamFeed"
-
-    def serialize(self, version, value, context):
-
-        context.write_uint(value[0])
-        data_type = value[2]
-        if version > 0:
-            context.write_byte(data_type)
-        if data_type == Stream.DataType.byte:
-            context.write_byte_array(value[1])
-        elif data_type == Stream.DataType.string:
-            context.write_using_serializer(self.__array, value[1])
-        else:
-            context.write_float_array(value[1])
-
-    def deserialize(self, version, context):
-
-        id = context.read_uint()
-        type = Stream.DataType.float
-        if version > 0:
-            type = Stream.DataType(context.read_byte())
-
-        if type == Stream.DataType.byte:
-            data = context.read_byte_array()
-        elif type == Stream.DataType.string:
-            data = context.read_using_serializer(self.__array)
-        else:
-            data = context.read_float_array()
-
-        return (id, data, type)
-
-
-class FeedStreamDone(serializer_fields.TypeSerializer):
-
-    def __init__(self):
-        pass
-
-    def version(self):
-        return 0
-
-    def name(self):
-        return "StreamFeedDone"
-
-    def serialize(self, version, value, context):
-        raise NotImplementedError
-
-    def deserialize(self, version, context):
-        return None
-
-
-class InterruptStream(serializer_fields.TypeSerializer):
-
-    def __init__(self):
-        pass
-
-    def version(self):
-        return 0
-
-    def name(self):
-        return "StreamInterrupt"
-
-    def serialize(self, version, value, context):
-        raise NotImplementedError
-
-    def deserialize(self, version, context):
-        err = context.read_byte()
-        id = context.read_uint()
-        return (err, id)
 
 
 class GetControllerTransforms(serializer_fields.TypeSerializer):
